requirements.txt
spotipy==2.11.1
Flask==1.1.2
python-dotenv==0.13.0
